telegram_bot/requirements.txt
pytelegramBotAPI
numpy
matplotlib
scikit-image
apache-tvm